# backend/src/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: PBCEx API
  version: 1.0.0
  description: |
    People's Bank & Commodities Exchange (PBCEx) API for precious metals trading, 
    custody, and fulfillment operations.

    ## Authentication
    All protected endpoints require a Bearer token in the Authorization header:
    ```
    Authorization: Bearer <your-access-token>
    ```

    ## Rate Limits
    - General API: 1000 requests per 15 minutes
    - Authentication: 5 attempts per 15 minutes  
    - KYC submissions: 10 per hour
    - Trading: 30 per minute

  contact:
    name: PBCEx API Support
    email: api-support@pbcex.com
  license:
    name: Proprietary
  x-build-info:
    branch: feature/money-movement-backend
    shortHash: ff3d03a
    generatedAt: '2025-08-31T20:00:00Z'
servers:
  - url: http://localhost:4000
    description: Development server
  - url: https://api.pbcex.com
    description: Production server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required:
        - code
        - message
        - timestamp
      properties:
        code:
          type: string
          enum:
            [
              VALIDATION_ERROR,
              AUTHENTICATION_ERROR,
              AUTHORIZATION_ERROR,
              NOT_FOUND,
              CONFLICT,
              RATE_LIMITED,
              INTERNAL_ERROR,
              SERVICE_UNAVAILABLE,
            ]
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
        details:
          type: object

    # Money Movement Schemas
...
    PaymentRequest:
      type: object
      properties:
        id:
          type: string
          format: uuid
        amount:
          type: string
        currency:
          type: string
        description:
          type: string
        status:
          type: string
        token:
          type: string
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /api/payment-requests:
    post:
      tags: [Payment Requests]
      operationId: createPaymentRequest
      summary: Create payment request
      description: Create a new payment request
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [asset, amount, expiresIn]
              properties:
                asset:
                  type: string
                  enum: [AU, AG, PT, PD, CU, USD, USDC]
                  description: Asset to request
                amount:
                  type: string
                  pattern: '^\d+\.?\d*$'
                  description: Amount to request
                expiresIn:
                  type: integer
                  minimum: 3600
                  maximum: 604800
                  description: Request expiration in seconds (1 hour to 1 week)
                memo:
                  type: string
                  description: Optional payment memo
                recipientEmail:
                  type: string
                  format: email
                  description: Recipient email address
      responses:
        '201':
          description: Payment request created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    example: 'SUCCESS'
                  data:
                    $ref: '#/components/schemas/PaymentRequest'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
      x-feature-flag: MONEY_MOVEMENT_ENABLED

    get:
      tags: [Payment Requests]
      operationId: getPaymentRequests
      summary: Get payment requests
      description: Retrieve user's payment requests
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Payment requests retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    example: 'SUCCESS'
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PaymentRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      x-feature-flag: MONEY_MOVEMENT_ENABLED
  /api/payment-requests/{id}:
    get:
      tags: [Payment Requests]
      operationId: getPaymentRequest
      summary: Get payment request details
      description: Retrieve specific payment request details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Payment request ID
      responses:
        '200':
          description: Payment request details retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    example: 'SUCCESS'
                  data:
                    $ref: '#/components/schemas/PaymentRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
      x-feature-flag: MONEY_MOVEMENT_ENABLED

    post:
      tags: [Payment Requests]
      operationId: cancelPaymentRequest
      summary: Cancel payment request
      description: Cancel an active payment request
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Payment request ID
      responses:
        '200':
          description: Payment request cancelled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    example: 'SUCCESS'
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: 'Payment request cancelled successfully'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
      x-feature-flag: MONEY_MOVEMENT_ENABLED
  /api/payment-requests/link/{token}:
    get:
      tags: [Payment Requests]
      operationId: getPaymentRequestByToken
      summary: Get payment request by token
      description: Retrieve payment request details using public token
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
          description: Payment request token
      responses:
        '200':
          description: Payment request details retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    example: 'SUCCESS'
                  data:
                    $ref: '#/components/schemas/PaymentRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/payment-requests/stats:
    get:
      tags: [Payment Requests]
      operationId: getPaymentRequestStats
      summary: Get payment request statistics
      description: Retrieve payment request statistics
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Payment request stats retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    example: 'SUCCESS'
                  data:
                    type: object
                    properties:
                      totalRequests:
                        type: integer
                      activeRequests:
                        type: integer
                      completedPayments:
                        type: integer
                      totalVolume:
                        type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
      x-feature-flag: MONEY_MOVEMENT_ENABLED

  # Recurring Endpoints
...